feat(vscode_ide): Code refactored
